+++ Firmware/MotorControl/controller.hpp
@@ -10,6 +10,8 @@ public:
     enum Error_t {
         ERROR_NONE = 0,
         ERROR_OVERSPEED = 0x01,
+        ERROR_INVALID_INPUT_MODE = 0x02,
+        ERROR_UNSTABLE_GAIN = 0x04,
     };
 
     // Note: these should be sorted from lowest level of control to
@@ -22,8 +24,17 @@ public:
         CTRL_MODE_TRAJECTORY_CONTROL = 4
     };
 
+    enum InputMode_t{
+        INPUT_MODE_INACTIVE,
+        INPUT_MODE_PASSTHROUGH,
+        INPUT_MODE_VEL_RAMP,
+        INPUT_MODE_POS_FILTER,
+        INPUT_MODE_MIX_CHANNELS,
+    };
+
     struct Config_t {
-        ControlMode_t control_mode = CTRL_MODE_POSITION_CONTROL;  //see: Motor_control_mode_t
+        ControlMode_t control_mode = CTRL_MODE_POSITION_CONTROL;  //see: ControlMode_t
+        InputMode_t input_mode = INPUT_MODE_INACTIVE;  //see: InputMode_t
         float pos_gain = 20.0f;  // [(counts/s) / counts]
         float vel_gain = 5.0f / 10000.0f;  // [A/(counts/s)]
         // float vel_gain = 5.0f / 200.0f, // [A/(rad/s)] <sensorless example>
@@ -32,6 +43,8 @@ public:
         float vel_limit_tolerance = 1.2f;  // ratio to vel_lim. 0.0f to disable
         float vel_ramp_rate = 10000.0f;  // [(counts/s) / s]
         bool setpoints_in_cpr = false;
+        float inertia = 0.0f;      // [A/(count/s^2)]
+        float input_filter_bandwidth = 2.0f; // [1/s]
     float homing_speed = 2000.0f;   // [counts/s]
     };
 
@@ -52,6 +65,7 @@ public:
     void start_anticogging_calibration();
     bool anticogging_calibration(float pos_estimate, float vel_estimate);
 
+    void update_filter_gains();
     bool update(float pos_estimate, float vel_estimate, float* current_setpoint);
 
     Config_t& config_;
@@ -87,8 +101,12 @@ public:
     // float vel_setpoint = 800.0f; <sensorless example>
     float vel_integrator_current_ = 0.0f;  // [A]
     float current_setpoint_ = 0.0f;        // [A]
-    float vel_ramp_target_ = 0.0f;
-    bool vel_ramp_enable_ = false;
+
+    float input_pos_ = 0.0f;
+    float input_vel_ = 0.0f;
+    float input_current_ = 0.0f;
+    float input_filter_kp_ = 0.0f;
+    float input_filter_ki_ = 0.0f;
 
     uint32_t traj_start_loop_count_ = 0;
 
@@ -96,14 +114,16 @@ public:
     auto make_protocol_definitions() {
         return make_protocol_member_list(
             make_protocol_property("error", &error_),
+            make_protocol_property("input_pos", &input_pos_),
+            make_protocol_property("input_vel", &input_vel_),
+            make_protocol_property("input_current", &input_current_),
             make_protocol_property("pos_setpoint", &pos_setpoint_),
             make_protocol_property("vel_setpoint", &vel_setpoint_),
             make_protocol_property("vel_integrator_current", &vel_integrator_current_),
             make_protocol_property("current_setpoint", &current_setpoint_),
-            make_protocol_property("vel_ramp_target", &vel_ramp_target_),
-            make_protocol_property("vel_ramp_enable", &vel_ramp_enable_),
             make_protocol_object("config",
                 make_protocol_property("control_mode", &config_.control_mode),
+                make_protocol_property("input_mode", &config_.input_mode),
                 make_protocol_property("pos_gain", &config_.pos_gain),
                 make_protocol_property("vel_gain", &config_.vel_gain),
                 make_protocol_property("vel_integrator_gain", &config_.vel_integrator_gain),
@@ -111,6 +131,9 @@ public:
                 make_protocol_property("vel_limit_tolerance", &config_.vel_limit_tolerance),
                 make_protocol_property("vel_ramp_rate", &config_.vel_ramp_rate),
                 make_protocol_property("setpoints_in_cpr", &config_.setpoints_in_cpr),
+                make_protocol_property("inertia", &config_.inertia),
+                make_protocol_property("input_filter_bandwidth", &config_.input_filter_bandwidth,
+                    [](void* ctx) { static_cast<Controller*>(ctx)->update_filter_gains(); }, this),
                 make_protocol_property("homing_speed", &config_.homing_speed)
             ),
             make_protocol_function("set_pos_setpoint", *this, &Controller::set_pos_setpoint,